refactor: refactor in pages and components for nomenclature correct
