+++ src/layouts/PetDexApp/components/SideMenu/index.scss
