+++ src/layouts/PetDexApp/pages/NoPetRegirested/index.scss
