+++ src/layouts/PetDexApp/pages/PetRegister/index.scss
